requirements.txt
sentence-transformers
faiss-cpu
fastapi
uvicorn
pandas
numpy
langdetect
google-generativeai
python-dotenv
